refactor: ♻️ (api) separate schema files
