+++ apps/api/drizzle.config.ts
@@ -2,11 +2,11 @@ import type { Config } from "drizzle-kit";
 
 export default {
 	dialect: "sqlite",
-	schema: "./src/infrastructure/drizzle/schema.ts",
+	schema: "./src/infrastructure/drizzle/schema/!(index).ts",
 	out: "./drizzle/migrations",
 	driver: "d1",
 	dbCredentials: {
 		wranglerConfigPath: "./wrangler.toml",
 		dbName: "mona-ca_db",
 	},
-} satisfies Config;
+} as const satisfies Config;
