+++ apps/api/src/infrastructure/drizzle/schema/index.ts
@@ -1,0 +1,6 @@
+export { oAuthAccounts } from "./oauth-accounts";
+export { passwordResetTokens } from "./password-reset-tokens";
+export { session } from "./session";
+export { users } from "./users";
+export { userRelations } from "./relations";
+export { emailVerificationCode } from "./email-verification-code";
